fix(buildPermissions): use bigint for erc20 periodic periodAmount
What: Set periodAmount to BigInt(usdc base units) per smart-accounts-kit types.
Why: TypeScript build rejected number where PermissionTypes expects bigint.
Impact: Activation permission request type-checks in production build.

+++ app/src/lib/delegation/buildPermissions.ts
@@ -15,8 +15,7 @@ export function buildActivationPermissions(
 ): PermissionRequestParameter[] {
   const monthlyCap = options?.monthlyCapUsdc ?? 500
   const tokenAddress = options?.usdcAddress ?? CONTRACTS.usdc
-  // MetaMask RPC cannot serialize BigInt in permission params.
-  const periodAmount = monthlyCap * 1_000_000
+  const periodAmount = BigInt(monthlyCap * 1_000_000)
 
   return [
     {
